move models to api.go file
