+++ api/api.go
@@ -7,12 +7,30 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/aviva-verde/tech-test-backend-go/models"
 	"github.com/gorilla/mux"
 )
 
 type Server struct {
 	Logger *slog.Logger
+}
+
+type Authority struct {
+	ID   int    `json:"id"`
+	Name string `json:"name"`
+}
+
+type AuthorityRating struct {
+	Name  string  `json:"name"`
+	Value float64 `json:"value"`
+}
+
+type FSAAuthorities struct {
+	Authorities []FSAAuthority `json:"authorities"`
+}
+
+type FSAAuthority struct {
+	ID   int    `json:"LocalAuthorityId"`
+	Name string `json:"Name"`
 }
 
 func (s Server) Start() {
@@ -45,15 +63,15 @@ func (s Server) getAuthorities(w http.ResponseWriter, r *http.Request) {
 	res, _ := http.DefaultClient.Do(req)
 	body, _ := io.ReadAll(res.Body)
 
-	var fsaAuthorities models.FSAAuthorities
+	var fsaAuthorities FSAAuthorities
 	err := json.Unmarshal(body, &fsaAuthorities)
 	if err != nil {
 		s.Logger.Error("error unmarshalling authorities", "error", err)
 	}
 
-	var authorities []models.Authority
+	var authorities []Authority
 	for _, authority := range fsaAuthorities.Authorities {
-		model := models.Authority(authority)
+		model := Authority(authority)
 		authorities = append(authorities, model)
 	}
 
@@ -65,7 +83,7 @@ func (s Server) getAuthorities(w http.ResponseWriter, r *http.Request) {
 
 func (s Server) getAuthority(w http.ResponseWriter, r *http.Request) {
 	s.Logger.Info("Getting authority")
-	authorityRating := []models.AuthorityRating{
+	authorityRating := []AuthorityRating{
 		{Name: "5-star", Value: 22.41},
 		{Name: "4-star", Value: 43.13},
 		{Name: "3-star", Value: 12.97},
